// backend/src/types/response/widget/getFloatingWidgetResponse.ts
export type GetFloatingWidgetResponse = {
  id: number;
  widget_color_header_background: string | null;
  widget_color_header_text: string | null;
  widget_color_text_title: string | null;
  widget_color_text_description: string | null;
  widget_color_button_background: string | null;
  widget_color_button_text: string | null;
  widget_color_widget_button_background: string | null;
  widget_color_widget_button_text: string | null;
  widget_color_others_link_color: string | null;
  widget_color_others_icon_color: string | null;
  widget_corner_shape_screen_launcher_edge: string | null;
  widget_corner_shape_screen_sections: string | null;
  widget_corner_shape_button: string | null;
  widget_corner_shape_text_fields: string | null;
  widget_banner_image: string | null;
  widget_brand_icon: string | null;
  widget_custom_css: string | null;
  widget_custom_icon_rewards: string | null;
  widget_custom_icon_ways_to_earn: string | null;
  widget_custom_icon_ways_to_redeem: string | null;
  widget_hide_widget_unless_deep_link_activated: number | null;
  widget_hide_widget_launcher_on_mobile: number | null;
  widget_brandlift_branding: number | null;
  widget_font_primary_font: string | null;
  widget_font_secondary_font: string | null;
  status: string | null;
  created_by: number | null;
  updated_by: number | null;
  user_id: number;
  admin_ref: number | null;
  created_at: Date;
  updated_at: Date | null;
};